started refactoring neighborhood module
